Move StreamedObject to its own place
