+++ src/rootfilespec/bootstrap/TList.py
@@ -1,5 +1,5 @@
-from rootfilespec.bootstrap.streamedobject import read_streamed_item
-from rootfilespec.bootstrap.TObject import StreamHeader, TObject, TObjectBits
+from rootfilespec.bootstrap.streamedobject import StreamHeader, read_streamed_item
+from rootfilespec.bootstrap.TObject import TObject, TObjectBits
 from rootfilespec.bootstrap.TString import TString
 from rootfilespec.dispatch import DICTIONARY
 from rootfilespec.serializable import serializable
