+++ src/rootfilespec/bootstrap/__init__.py
@@ -17,12 +17,13 @@ from rootfilespec.bootstrap.array import (
 from rootfilespec.bootstrap.assumed import ROOT3a3aTIOFeatures, TAtt3D, TVirtualIndex
 from rootfilespec.bootstrap.compression import RCompressed, RCompressionHeader
 from rootfilespec.bootstrap.RAnchor import ROOT3a3aRNTuple
+from rootfilespec.bootstrap.streamedobject import StreamedObject
 from rootfilespec.bootstrap.TBasket import TBasket
 from rootfilespec.bootstrap.TDirectory import TDirectory, TKeyList
 from rootfilespec.bootstrap.TFile import ROOTFile, TFile
 from rootfilespec.bootstrap.TKey import TKey
 from rootfilespec.bootstrap.TList import TList, TObjArray
-from rootfilespec.bootstrap.TObject import StreamedObject, TNamed, TObject
+from rootfilespec.bootstrap.TObject import TNamed, TObject
 from rootfilespec.bootstrap.TStreamerInfo import (
     TStreamerBase,
     TStreamerElement,
